Release 0.1.5 with C++20 support (#10)
* Release 0.1.5 with C++20 support

* Use current LLVM for macOS CI

* Support the system AppleClang C++20 library

* Organize examples and tests by target

---------

+++ ruvia-web/src/auth/JwtJson.cpp
@@ -57,7 +57,7 @@ jwtParseJsonNumericDate(std::string_view value) {
         }
     }
 
-    long double fractional = 0;
+    double fractional = 0;
     const auto [fractionalPtr, fractionalEc] = std::from_chars(
         value.data(), value.data() + value.size(), fractional);
     if (fractionalEc != std::errc{} || !std::isfinite(fractional)) {
